move functions that do more than wrap a command to git_util.py
