+++ varats-core/varats/utils/git_util.py
@@ -7,10 +7,9 @@ from itertools import chain
 from pathlib import Path
 
 import pygit2
-from benchbuild.utils.cmd import git
+from benchbuild.utils.cmd import git, grep
 from benchbuild.utils.revision_ranges import RevisionRange
-from plumbum import local
-from plumbum.commands.modifiers import RETCODE
+from plumbum import local, TF, RETCODE
 
 from varats.project.project_util import (
     get_local_project_gits,
@@ -914,3 +913,34 @@ class RevisionBinaryMap(tp.Container[str]):
                     return True
 
         return False
+
+
+def has_branch(repo_folder: Path, branch_name: str) -> bool:
+    """Checks if a branch exists in the local repository."""
+
+    exit_code = git["-C",
+                    repo_folder.absolute(), "rev-parse", "--verify",
+                    branch_name] & TF
+    return tp.cast(bool, exit_code)
+
+
+def has_remote_branch(repo_folder: Path, branch_name: str, remote: str) -> bool:
+    """Checks if a remote branch of a repository exists."""
+    exit_code = (
+        git["-C",
+            repo_folder.absolute(), "ls-remote", "--heads", remote, branch_name]
+        | grep[branch_name]
+    ) & RETCODE
+    return tp.cast(bool, exit_code == 0)
+
+
+def branch_has_upstream(
+    repo_folder: Path, branch_name: str, upstream: str = 'origin'
+) -> bool:
+    """Check if a branch has an upstream remote."""
+    exit_code = (
+        git["-C",
+            repo_folder.absolute(), "rev-parse", "--abbrev-ref",
+            branch_name + "@{upstream}"] | grep[upstream]
+    ) & RETCODE
+    return tp.cast(bool, exit_code == 0)
